+++ src/assets/froala/FroalaAsset.php
@@ -28,6 +28,45 @@ class FroalaAsset extends AssetBundle
         'selectAll'       => 'Select All',
     ];
 
+    const CORE_LANGUAGES = [
+        'ar'    => 'Arabic',
+        'bs'    => 'Bosnian',
+        'cs'    => 'Czech',
+        'da'    => 'Danish',
+        'de'    => 'German',
+        'en_ca' => 'English Canada',
+        'en_gb' => 'English United Kingdom',
+        'es'    => 'Spanish',
+        'et'    => 'Estionian',
+        'fa'    => 'Persian',
+        'fi'    => 'Finish',
+        'fr'    => 'French',
+        'he'    => 'Hebrew',
+        'hr'    => 'Croatian',
+        'hu'    => 'Hungarian',
+        'id'    => 'Idonesian',
+        'it'    => 'Italian',
+        'ja'    => 'Japanese',
+        'ko'    => 'Korean',
+        'me'    => 'Montenegrin',
+        'nb'    => 'Norwegian',
+        'nl'    => 'Dutch',
+        'pl'    => 'Polish',
+        'pt_br' => 'Portuguese Brazil',
+        'pt_pt' => 'Portuguese Portugal',
+        'ro'    => 'Romanian',
+        'ru'    => 'Russian',
+        'sk'    => 'Slovak',
+        'sr'    => 'Serbian',
+        'sv'    => 'Swedish',
+        'th'    => 'Thai',
+        'tr'    => 'Turkish',
+        'uk'    => 'Ukrainian',
+        'vi'    => 'Vietnamese',
+        'zh_cn' => 'Chinese China',
+        'zh_tw' => 'Chinese Taiwan',
+    ];
+
     /**
      * @var array
      */
@@ -53,6 +92,12 @@ class FroalaAsset extends AssetBundle
         $this->js = [
             'js/froala_editor.pkgd.min.js',
         ];
+
+        // add-in language
+        $language = self::getLanguage();
+        if (array_key_exists(strtolower($language), self::CORE_LANGUAGES)) {
+            $this->js = array_merge($this->js, [sprintf('js/languages/%s.js', $language)]);
+        }
 
         parent::init();
     }
@@ -102,5 +147,27 @@ class FroalaAsset extends AssetBundle
         }
 
         return $this->loadedPlugins;
+    }
+
+    /**
+     * Returns the language string applicable to the Froala Editor.
+     *
+     * @return string
+     */
+    public static function getLanguage()
+    {
+        $craftLanguage = Craft::$app->getTargetLanguage();
+        $craftLanguage = $language = strtolower(str_replace('-', '_', $craftLanguage));
+
+        if (!array_key_exists($language, self::CORE_LANGUAGES)) {
+            $language = substr($language, 0, strpos($language, '_'));
+        }
+
+        if (array_key_exists(strtolower($language), self::CORE_LANGUAGES)) {
+
+            return $language;
+        }
+
+        return $craftLanguage;
     }
 }